[Toolbox] Cleanup local toolbox
+ Add devtools-local-environment
+ Add devtools-client-adapters

The goal is to focus the local toolbox on *just* showing the Landing
page and establishing the connection.

The local environment is responsible for scripts related to working
locally. Examples are
* a webpack config for local development and bundling
* a dev server
* running unit tests
* running mochi tests
* bundling

The Client adapters is responsible for connecting to the browser
as well as providing a consitent API for the debugger.